M_PI could be a double on some systems so cast it as a float when it's used as a float
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18018 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/CustomWorld.cxx
@@ -58,9 +58,9 @@ void CustomWorld::writeToWorld(WorldInfo* world) const
   if (createWalls) {
     float wallHeight = BZDB.eval(StateDatabase::BZDB_WALLHEIGHT);
     float worldSize = BZDBCache::worldSize * 0.5f;
-    world->addWall(0.0f, worldSize, 0.0f, 1.5f*M_PI, worldSize, wallHeight);
-    world->addWall(worldSize, 0.0f, 0.0f, M_PI, worldSize, wallHeight);
-    world->addWall(0.0f, -worldSize, 0.0f, 0.5f*M_PI, worldSize, wallHeight);
+    world->addWall(0.0f, worldSize, 0.0f, 1.5f*(float)M_PI, worldSize, wallHeight);
+    world->addWall(worldSize, 0.0f, 0.0f, (float)M_PI, worldSize, wallHeight);
+    world->addWall(0.0f, -worldSize, 0.0f, 0.5f*(float)M_PI, worldSize, wallHeight);
     world->addWall(-worldSize, 0.0f, 0.0f, 0.0f, worldSize, wallHeight);
   }
 }
